fix: allow path parameters
