+++ components/script/methods/get/index.js
@@ -1,11 +1,6 @@
 module.exports = ({type}) => {
     return async function get(msg, $meta) {
-        const cryptoData = await this.bus.importMethod('crypto.record.getBulk')(msg.id.split(','));
-        return cryptoData.records.map(record => {
-            return {
-                id: record.id,
-                data: record.data
-            };
-        });
+        const {id, data} = await this.bus.importMethod('crypto.record.get')({id: msg.id});
+        return {id, data};
     };
 };
