+++ package.json
@@ -32,7 +32,7 @@
     "ut-port-performance": "^6.0.0",
     "ut-port-script": "^6.0.0",
     "ut-port-sql": "^7.3.1",
-    "ut-port-swagger": "^1.2.2",
+    "ut-port-swagger": "^1.3.0",
     "ut-run": "^9.7.0-rc-einstein.5"
   },
   "devDependencies": {
